+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - General
   - Migrate project from depstar -> tools.build. #711
+  - Improve classpath hash to consider invaliding cache when local root deps was changed.
   
 - Editor
   - Support going to namespace definition on an alias. #706